@@ -34,66 +34,55 @@
   },
   "homepage": "https://github.com/react-bootstrap/react-router-bootstrap",
   "peerDependencies": {
-    "react": ">=0.14.0",
-    "react-router-dom": ">=4.0.0"
+    "react": ">=16.13.1",
+    "react-router-dom": ">=6.0.0"
   },
   "devDependencies": {
-    "babel-cli": "^6.11.4",
-    "babel-core": "^6.11.4",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.2.4",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-es2015-loose": "^7.0.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-1": "^6.5.0",
-    "babel-register": "^6.11.5",
-    "bootstrap": "^3.3.7",
-    "chai": "^3.5.0",
-    "colors": "^1.1.2",
-    "css-loader": "^0.23.1",
-    "eslint": "^2.13.1",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-babel": "^3.3.0",
-    "eslint-plugin-import": "^1.11.1",
-    "eslint-plugin-jsx-a11y": "^1.5.5",
-    "eslint-plugin-mocha": "^4.1.0",
-    "eslint-plugin-react": "^5.2.2",
-    "file-loader": "^0.9.0",
-    "history": "^2.1.2",
-    "html-webpack-plugin": "^2.22.0",
-    "karma": "^1.1.1",
-    "karma-chrome-launcher": "^1.0.1",
-    "karma-firefox-launcher": "^1.0.0",
-    "karma-mocha": "^1.1.1",
-    "karma-mocha-reporter": "^2.0.4",
-    "karma-sinon-chai": "^1.2.3",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^1.7.0",
-    "less": "^2.7.1",
-    "less-loader": "^2.2.3",
-    "lodash": "^4.14.0",
-    "mocha": "^2.5.3",
-    "react": "^15.2.1",
-    "react-addons-test-utils": "^15.4.2",
-    "react-bootstrap": "^0.30.0",
-    "react-dom": "^15.2.1",
-    "react-router-dom": "^4.0.0",
-    "release-script": "^1.0.2",
-    "rimraf": "^2.5.4",
-    "shelljs": "^0.7.2",
-    "sinon": "^1.17.4",
-    "sinon-chai": "^2.8.0",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.1",
-    "webpack-dev-server": "^1.14.1",
-    "yargs": "^4.8.1"
-  },
-  "release-script": {
-    "bowerRepo": "git@github.com:react-bootstrap/react-router-bootstrap-bower.git"
+    "@babel/core": "^7.11.4",
+    "@babel/preset-env": "^7.11.0",
+    "@babel/preset-react": "^7.10.4",
+    "babel-eslint": "^10.1.0",
+    "babel-jest": "^26.3.0",
+    "babel-loader": "^8.1.0",
+    "bootstrap": "^4.5.2",
+    "chai": "^4.2.0",
+    "colors": "^1.4.0",
+    "css-loader": "^4.2.2",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.3",
+    "eslint": "^7.7.0",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-plugin-babel": "^5.3.1",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
+    "eslint-plugin-react": "^7.20.6",
+    "file-loader": "^6.0.0",
+    "history": "^5.0.0",
+    "html-webpack-plugin": "^4.3.0",
+    "jasmine-enzyme": "^7.1.2",
+    "jest": "^26.4.2",
+    "lodash": "^4.17.20",
+    "react": "^16.13.1",
+    "react-bootstrap": "^1.3.0",
+    "react-dom": "^16.13.1",
+    "react-router-dom": "^6.0.0-beta.0",
+    "rimraf": "^3.0.2",
+    "style-loader": "^1.2.1",
+    "url-loader": "^4.1.0",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12",
+    "webpack-dev-server": "^3.11.0",
+    "yargs": "^15.4.1"
   },
   "dependencies": {
-    "prop-types": "^15.5.10"
+    "prop-types": "^15.7.2"
+  },
+  "jest": {
+    "setupFilesAfterEnv": [
+      "<rootDir>test/setupTests.js"
+    ],
+    "transform": {
+      "^.+\\.js$": "babel-jest"
+    }
   }
 }
