@@ -1,6 +1,6 @@
 import React from 'react';
-import Nav from 'react-bootstrap/lib/Nav';
-import NavItem from 'react-bootstrap/lib/NavItem';
+import Nav from 'react-bootstrap/Nav';
+import NavItem from 'react-bootstrap/NavItem';
 import { Link } from 'react-router-dom';
 
 import LinkContainer from '../../src/LinkContainer';
@@ -11,14 +11,14 @@ export default () => (
     <h2>NavItem</h2>
 
     <h3>Baseline</h3>
-    <Nav bsStyle="pills" activeKey={1}>
-      <NavItem eventKey={1}>NavItem 1 content</NavItem>
-      <NavItem eventKey={2}>NavItem 2 content</NavItem>
-      <NavItem eventKey={3} disabled>NavItem 3 content</NavItem>
+    <Nav variant="pills" activeKey={1}>
+      <NavItem>NavItem 1 content</NavItem>
+      <NavItem>NavItem 2 content</NavItem>
+      <NavItem disabled>NavItem 3 content</NavItem>
     </Nav>
 
     <h3>LinkContainer</h3>
-    <Nav bsStyle="pills">
+    <Nav variant="pills">
       <LinkContainer to="/nav-item">
         <NavItem>NavItem 1 content</NavItem>
       </LinkContainer>
