@@ -6,7 +6,7 @@ export default {
   module: {
     loaders: [
       { test: /\.js/, loader: 'babel', exclude: /node_modules/ },
-      { test: /\.less$/, loader: 'style!css!less' },
+      { test: /\.css$/, loader: 'style!css' },
       { test: /\.woff|\.woff2$/, loader: 'url?prefix=font/&limit=5000' },
       { test: /\.eot$|\.ttf$|\.svg$/, loader: 'file?prefix=font/' },
     ],
